refactor: ReservationWithWaitingOrder를 domain으로 이동
