+++ src/main/java/roomescape/repository/JdbcReservationRepository.java
@@ -16,7 +16,7 @@ import roomescape.domain.ReservationStatus;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
 import roomescape.domain.WaitingOrder;
-import roomescape.service.dto.ReservationWithWaitingOrder;
+import roomescape.domain.ReservationWithWaitingOrder;
 
 @Repository
 public class JdbcReservationRepository implements ReservationRepository {
