+++ src/main/java/roomescape/service/dto/ReservationResult.java
@@ -2,6 +2,7 @@ package roomescape.service.dto;
 
 import java.time.LocalDate;
 import roomescape.domain.ReservationStatus;
+import roomescape.domain.ReservationWithWaitingOrder;
 import roomescape.domain.WaitingOrder;
 
 public record ReservationResult(
